--- agent/version/version.go
@@ -1,3 +1,0 @@
-package version
-
-var Version string = "v0.0.1"
