+++ operator/agent/version/version.go
@@ -1,0 +1,3 @@
+package version
+
+var Version string = "v0.0.1"
